Refactory for ProfileService in Identity project
